refactor: ProductApiController(API 컨트롤러)와 ProductController(SSR) 분리
